ai-server/requirements.txt
fastapi
uvicorn
openai
langchain
langchain-openai
python-multipart
python-dotenv
